requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
web3>=7.0.0
requests>=2.31.0
